Communication established between client and server
A connection is established between the client and the server, sending a Heartbeat message across the network.  A NetUtility class was created to be used by both the client and the server.

The Message Base Class was populated with a Serialise and Deserialise function as well as functions for the Client and Server.

A Heartbeat message class was created with no payload to communicate between the client and server

The UI was hooked up to use these systems, and communication happens between two instances of the game